fix(telemetry): enforce parity review invariants
